Updated Naming Conventions
- Updated the vectorstore service naming conventions
- Updated the keywords service naming conventions
