+++ src/api/vectorstore/vectorstore.controller.ts
@@ -3,8 +3,8 @@ import fs from 'fs';
 import path from 'path';
 import { ResponseHandler } from '../../common/handlers/response.handler';
 import { ErrorHandler } from '../../common/handlers/error.handler';
-import { VectorstoreService } from '../../modules/vectorstores/vectorstore.service';
-import { KeywordService } from '../../modules/vectorstores/keywords.service';
+import { VectorstoreService } from '../../modules/vectorstore/vectorstore.service';
+import { KeywordService } from '../../modules/vectorstore/keywords.service';
 import { VectorStoreValidator } from './vectorstore.validator';
 import { Loader } from '../../startup/loader';
 import { DocumentProcessor } from '../../modules/document.processors/document.processor';
