+++ src/modules/module.injector.ts
@@ -3,7 +3,8 @@ import { DependencyContainer } from 'tsyringe';
 // import { CommunicationInjector } from './communication/communication.injector';
 // import { EhrInjector } from './ehr/ehr.injector';
 import { FileStorageInjector } from './storage/file.storage.injector';
-import { VectorstoreInjector, KeywordInjector } from './vectorstores/vectorstores.injector';
+import { VectorstoreInjector } from './vectorstore/vectorstore.injector';
+import { KeywordsInjector } from './vectorstore/keywords.injector';
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -15,7 +16,7 @@ export class ModuleInjector {
         // CommunicationInjector.registerInjections(container);
         FileStorageInjector.registerInjections(container);
         VectorstoreInjector.registerInjections(container);
-        KeywordInjector.registerInjection(container);
+        KeywordsInjector.registerInjection(container);
 
     }
