+++ src/modules/vectorstore/interfaces/keywords.service.interface.ts
@@ -1,0 +1,8 @@
+export interface IKeywordsService {
+    connectKeywordVectorStore();
+    disconnectKeywordVectorStore();
+    addKeywords(tenantId: string, tags: string[], fileName: string): Promise<string>;
+    updateKeywords();
+    deleteKeywords();
+    searchKeywords(tenantId: string, userQuery: string);
+}
