+++ src/modules/vectorstore/interfaces/vectorstore.service.interface.ts
@@ -10,12 +10,3 @@ export interface IVectorStoreService {
     loadVectorStore(clientName: string, projectName: string, collectionName: string);
     similaritySearch(tenantId: string, userQuery: string);
 }
-
-export interface IKeywordService {
-    connectKeywordVectorStore();
-    disconnectKeywordVectorStore();
-    addKeywords(tenantId: string, tags: string[], fileName: string): Promise<string>;
-    updateKeywords();
-    deleteKeywords();
-    searchKeywords(tenantId: string, userQuery: string);
-}
